--- src/components/card.vue
@@ -1,57 +1,0 @@
-<template>
-    <div class="card" v-if="isCardVisible">
-        <div class="card-header">
-            <h3 class="card-title">{{ title }}</h3>
-    
-            <slot name="card-tools">
-                <div class="card-tools">
-                    <span class="badge badge-danger" v-if="badge">{{ badge }}</span>
-                    <button type="button" class="btn btn-tool" @click="toggleCardBody" v-if="toolsBtn === 'on'">
-                        <i :class="isCardBodyVisible ? 'fas fa-minus' : 'fas fa-plus'"></i>
-                    </button>
-                    <button type="button" class="btn btn-tool" @click="closeCard" v-if="toolsBtn ==='on'">
-                        <i class="fas fa-times"></i>
-                    </button>
-                </div>        
-            </slot>
-        </div>
-        <!-- /.card-header -->
-        <div class="card-body p-0" v-show="isCardBodyVisible">
-            <slot></slot>
-        </div>
-        <!-- /.card-body -->
-        <div class="card-footer text-center" v-if="$slots.footer" v-show="isCardBodyVisible">
-            <slot name="footer"></slot>
-        </div>
-        <!-- /.card-footer -->
-    </div>
-</template>
-    
-<script setup lang="ts">
-import { ref, defineProps, PropType } from 'vue';
-const props = defineProps({
-    title: {
-        type: String as PropType<string>,
-        required: true
-    },
-    badge: {
-        type: String as PropType<string>,
-        default: ''
-    },
-    toolsBtn: {
-        type: String as PropType<string>,
-        default: 'on'
-    }
-});
-
-const isCardVisible = ref(true);
-const isCardBodyVisible = ref(true);
-
-const toggleCardBody = () => {
-    isCardBodyVisible.value = !isCardBodyVisible.value;
-}
-
-const closeCard = () => {
-    isCardVisible.value = false;
-}
-</script>
