+++ src/components/field/index.ts
@@ -3,3 +3,4 @@ export { default as NixDropDown } from './NixDropDown.vue';
 export { default as NixOptionGroup } from './NixOptionGroup.vue';
 export { default as NixDatePicker } from './NixDatePicker.vue';
 export { default as NixCheckBox } from './NixCheckBox.vue';
+export { default as DualListbox } from './dual-listbox.vue';
