+++ src/components/index.ts
@@ -1,9 +1,3 @@
-export { default as DataGrid } from './DataGrid.vue';
-export { default as Attendance } from './Attendance.vue';
-export { default as Card } from './card.vue';
-export { default as DualListbox } from './dual-listbox.vue';
-export { default as Footer } from './Footer.vue';
-
 export * from './datagrid/index';
 export * from './layout/index';
 export * from './card/index';
