+++ src/router/index.ts
@@ -5,7 +5,6 @@ import Dashboard from '@/pages/dashboard.vue';
 import Dashboard2 from '@/pages/dashboard2.vue';
 import Maps from '@pages/Maps.vue';
 import Tables from '@pages/Tables.vue';
-import Tables2 from '@pages/Tables2.vue';
 import Calendars from '@pages/Calendars.vue';
 import Modals from '@pages/Modals.vue'
 import DatePickers from '@/pages/DatePickers.vue';
@@ -42,10 +41,6 @@ const routes: Array<RouteRecordRaw> = [
             {
                 path: 'tables',
                 component: Tables
-            },
-            {
-                path: 'tables2',
-                component: Tables2
             },
             {
                 path: 'calendars',
